Create Scale_intensity_check_clip.txt
Opens all sounds in a directory, scales them to a uniform intensity, checks to see if they would clip. If they clip, a window opens up asking you if you want to lower the output intensity in an attempt to avoid clipping. offers the option to save after sounds are checked. 
NOTE: this script is a lot slower than the simpler scale-and-save version.