Create Make_duration_continuum.txt
Start with one or multiple sounds. If using multiple sounds, ensure that they all have the same timing structure. The first selected sound is called up for you to select the onset and offset of the duration manipulation (so that you can manipulate just the vowel of a CVC, for instance). 

You can use just one sound and make a single duration continuum, or use many sounds that are the product of a different script (for example, the formant manipulation script), and make continua for all of them at once.